testing selex disabling
Rick asked to confirm that the two appraoches are identical if movement is off and F equal. My prelim testing said no, which suggests that even in the absence of movement fish were being unequally targeted among regions, which would happen with the distinct selectivity regimes. Somewhat confounded with growth obviously, but building this in as a test.