include F mort
recall that f mort is really more a harvest rate here, so we take out 1-FF[area]
also, for the calcs to work we assign the F in area to enact on individuals in area before they move. Thus NCOME is the total who survived natural mortality in area j and fishing mortality in area j and arrive in area i, just as area i fish stay put, survive and are fished.