test & prop for bivariate uniroot
first got some strange banding, and values were lower than expected. we would think the uniroot would give us a total F and prop equal to the sum of X Y on the yield heatmap and X/(X+Y). At first it wasn't and it was because i accidentally used the same F for both areas in the second part of the delta