build doPR which runs 100+ iters w hybrid method
this works so long as you divide by y when saving the final NPR. This is basically saying that the NPR in year 100 is relative to 100 recruits, which is how much the system has seen by that point. The figures saved here were NOT using the hybrid method however, only confirmed the shape of the stable age curve visually.