sb0 should not change (?)
noticed that SB0 varies with F due to the fact that Req varies with F. Though tmp0$SBPR is equivalent among methods and doesn't care about F (it's unfished), the variation in Req was causing this to decline and upsetting depletion. SB_0 = SBPR0*R0, split by putative proportion, no matter what. this also matters because SB0, in addiiton to RProp input and R0 input are placed inside the SRR to get expected R.