code up three options for doing NAA2
one is a cheap version taht can't accomodate a symmmove setup.
the other two are the same except the first one fishes BEFORE (or coincident to) movement, and will return the conservation of FMSY result we expected to see. The final one is what AEP and I discussed verbally, but it doesn't behave the same way because mortality is effectively reduced in the source area so FMSY_source is inflated. I am suggesting to not run with this version since it presents quite different results.